`rpc` testing utilities are only available in tests
